+++ src/main/java/open/commons/spring/web/autoconfigure/OpenCommonsSpringWebCoreAutoConfiguration.java
@@ -43,7 +43,7 @@ import open.commons.spring.web.configure.concurrent.executor.ScheduledExecutorCo
 import open.commons.spring.web.configure.concurrent.task.TaskExecutorConfiguration;
 import open.commons.spring.web.configure.concurrent.task.TaskSchedulerConfiguration;
 import open.commons.spring.web.configure.exception.ExceptionHttpStatusBinderConfiguration;
-import open.commons.spring.web.configure.jdbc.DataSourceUrlProcessor;
+import open.commons.spring.web.configure.jdbc.DataSourcePostProcessor;
 import open.commons.spring.web.configure.resttemplate.RestTemplateConfiguration;
 import open.commons.spring.web.environment.resolve.EnvironmentResolver;
 import open.commons.spring.web.environment.resolve.strategy.DockerHostGatewayStrategy;
@@ -117,7 +117,7 @@ import open.commons.spring.web.environment.resolve.strategy.DockerHostGatewayStr
         , RestTemplateConfiguration.class //
         , ServletConfiguration.class //
         , DockerHostGatewayStrategy.class //
-        , DataSourceUrlProcessor.class //
+        , DataSourcePostProcessor.class //
         , EnvironmentResolver.class //
 })
 public class OpenCommonsSpringWebCoreAutoConfiguration {