+++ history.md
@@ -1,7 +1,19 @@
+[2026/07/10]
+- 추가
+  + 'docker 실행환경'에서 `localhost`, `127.0.0.1` 값에 대한 `host-gateway` 지원 기능 추가
+    + open.commons.spring.web.configure.jdbc
+      + DataSourcePostProcessor
+      + DataSourceUrlProperties
+    + open.commons.spring.web.environment.resolve
+      + EnvironmentResolver
+    + open.commons.spring.web.environment.resolve.strategy
+      + DockerHostGatewayStrategy
+      + ResolveStrategy
+     
 [2026/07/09]
 - 추가
-  - resource/META-INF/open-commons/open-commons-spring-web.yml
-    - `open-commons.docker.extra_hosts.host_gateway`: 'docker bridge' 네트워크의 Gateway IP 를 갖는 '도메인' 정보.
+  + resource/META-INF/open-commons/open-commons-spring-web.yml
+    + `open-commons.docker.extra_hosts.host_gateway`: 'docker bridge' 네트워크의 Gateway IP 를 갖는 '도메인' 정보.
 
 [2026/06/30]
 - 추가